refactor model parser
