+++ packages/core/src/types/_common.ts
@@ -1,4 +1,4 @@
-import type { Tagged } from 'type-fest'
+import type { Simplify, Tagged, UnionToIntersection } from 'type-fest'
 
 export type NonEmptyArray<T> = [T, ...T[]]
 
@@ -16,3 +16,41 @@ export interface XYPoint {
 }
 
 export type KeysOf<T> = keyof T extends infer K extends string ? K : never
+
+type AllNever<Expressions> = UnionToIntersection<
+  {
+    [Name in keyof Expressions]: {
+      -readonly [Key in keyof Expressions[Name]]?: never
+    }
+  }[keyof Expressions]
+>
+
+/**
+ * @example
+ * ```
+ *   type Variant1 = {
+ *     a: string
+ *     b: number
+ *   }
+ *   type Variant2 = {
+ *     a: boolean
+ *   }
+ *
+ *   type Variants = ExclusiveUnion<{
+ *     Variant1: Variant1,
+ *     Variant2: Variant2
+ *   }>
+ *
+ *   // Type 'true' is not assignable to type 'string'.
+ *   const variant1: Variants = {
+ *      a: true,
+ *      b: 1
+ *   }
+ *
+ * ```
+ */
+
+export type ExclusiveUnion<Expressions> = Expressions extends object ? {
+    [Name in keyof Expressions]: Simplify<Omit<AllNever<Expressions>, keyof Expressions[Name]> & Expressions[Name]>
+  }[keyof Expressions]
+  : Expressions
