+++ packages/core/src/types/index.ts
@@ -2,6 +2,7 @@ export * from './_common'
 export * from './deployments'
 export * from './element'
 export * from './expression'
+export * from './fqn-ref'
 export * from './global'
 export * from './model'
 export * from './operators'
