+++ packages/language-server/src/__tests__/deployment-views.spec.ts
@@ -90,7 +90,7 @@ async function mkTestServices({ expect }: TestContext) {
 
   const validateRules = (rules: string) =>
     validateView(`
-      view {
+      deployment view view {
         ${rules}
       }
     `)
