+++ packages/language-server/src/validation/_shared.ts
@@ -8,7 +8,6 @@ export const RESERVED_WORDS = [
   'it',
   'self',
   'super',
-  'instance',
   'likec4lib',
   'global'
 ]
